* commit 'adf5e63aca6dd4c77b0baa785e95e3e739018a72':
  Leave more space for extensions in pthread_rwlockattr_t

+++ libc/include/pthread.h
@@ -91,7 +91,7 @@ typedef struct {
 typedef long pthread_mutexattr_t;
 typedef long pthread_condattr_t;
 
-typedef int pthread_rwlockattr_t;
+typedef long pthread_rwlockattr_t;
 
 typedef struct {
   pthread_mutex_t lock;
